Force f64 precision. Two work arounds: (1) in context.py, allow builtins called from python to be accepted as float32 or float64 (whichever is expected), (2) in quat.h and spatial.h, manually set default return type for quat_identity and transform_identity to quatd and transformd in python context

+++ warp/native/quat.h
@@ -1223,7 +1223,7 @@ inline CUDA_CALLABLE mat_t<Rows,Cols,Type>::mat_t(const vec_t<3,Type>& pos, cons
     data[3][3] = Type(1);
 }
 
-template<typename Type=float32>
+template<typename Type=float64>
 inline CUDA_CALLABLE quat_t<Type> quat_identity()
 {
     return quat_t<Type>(Type(0), Type(0), Type(0), Type(1));
